changing everything from f32 to f64

+++ src/constants.rs
@@ -1,5 +1,5 @@
 use nalgebra::{Matrix1x4};
-pub type color = [f32; 3];
+pub type color = [f64; 3];
 
 pub struct Canvas {
     pub pixels: Vec<Vec<color>>,
@@ -16,11 +16,11 @@ impl Canvas {
     }
 }
 pub struct _Env {
-    pub gravity: Matrix1x4<f32>,
-    pub wind: Matrix1x4<f32>
+    pub gravity: Matrix1x4<f64>,
+    pub wind: Matrix1x4<f64>
 }
 impl _Env {
-    pub fn new(g: f32, w: f32) -> Self {
+    pub fn new(g: f64, w: f64) -> Self {
         Self {
             gravity: Matrix1x4::new(0.0, -g, 0.0, 0.0),
             wind: Matrix1x4::new(-w, 0.0, 0.0, 0.0)
