+++ src/light.rs
@@ -1,12 +1,12 @@
 use nalgebra::Matrix4x1;
 
 pub struct PointLight {
-    pub intensity :f32,
-    pub position: Matrix4x1<f32>
+    pub intensity :f64,
+    pub position: Matrix4x1<f64>
 }
 
 impl PointLight {
-    pub fn new(intensity: f32, position: Matrix4x1<f32>) -> Self {
+    pub fn new(intensity: f64, position: Matrix4x1<f64>) -> Self {
         Self {
             intensity,
             position
